% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jqScrollBar.R
\name{jqScrollBar}
\alias{jqScrollBar}
\title{jqScrollBar}
\usage{
jqScrollBar(inputId, choices = choices, selected = null)
}
\arguments{
\item{inputId}{the id of this shiny input}

\item{value}{the initial value for this control}
}
\description{
jqScrollBar
}
