% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jqScrollBar.R
\name{updateJqScrollBar}
\alias{updateJqScrollBar}
\title{updateJqScrollBar
server to client update}
\usage{
updateJqScrollBar(session, inputId, cmd, value = "bogus")
}
\arguments{
\item{session}{the shiny session}

\item{inputId}{the control Id}

\item{value}{update with this value}
}
\description{
updateJqScrollBar
server to client update
}
